backend/pyproject.toml
[tool.poetry]
name = "backend-api"
version = "0.1.0"
description = "FastAPI Backend with Hexagonal Architecture"
authors = ["Army Dev <dev@armydev.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13.7"
fastapi = {extras = ["standard"], version = "^0.116.1"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.43"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
redis = "^5.2.0"
greenlet = "^3.1.1"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.18"
pyjwt = "^2.8.0"
bcrypt = "^4.0.1"
qrcode = {extras = ["pil"], version = "^7.4.2"}
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
strict = true